Extract test helper for setup a sheet session.
